+++ uni_modules/sky-nutui/components/sky-nutui/packages/__VUE/noticebar/index.vue
@@ -11,10 +11,11 @@
       <view class="left-icon" v-if="iconShow" :style="{ 'background-image': `url(${iconBg})` }">
         <slot name="left-icon"><nut-icon name="notice" size="16" :color="color" v-if="!iconBg"></nut-icon></slot>
       </view>
-      <view ref="wrap" class="wrap">
+      <view ref="wrap" :class="`wrap wrap${id}`">
         <view
           ref="content"
-          :class="['content', animationClass, { 'nut-ellipsis': isEllipsis }]"
+          class="content"
+          :class="[animationClass, { 'nut-ellipsis': isEllipsis }, `content${id}`]"
           :style="contentStyle"
           @animationend="onAnimationEnd"
           @webkitAnimationEnd="onAnimationEnd"
@@ -30,12 +31,13 @@
     <view class="nut-noticebar-vertical" v-if="scrollList.length > 0 && direction == 'vertical'" :style="barStyle">
       <template v-if="slots.default">
         <view class="horseLamp_list" :style="horseLampStyle">
-          <ScrollItem
+          <view
             v-for="(item, index) in scrollList"
-            :key="index"
+            v-bind:key="index"
             :style="{ height: height + 'px', 'line-height': height + 'px' }"
-            :item="item"
-          ></ScrollItem>
+          >
+          {{item}}
+          </view>
         </view>
       </template>
 
@@ -76,26 +78,13 @@ import {
   onDeactivated,
   ref,
   watch,
-  h
+  h,
+  getCurrentInstance
 } from 'vue';
 import { createComponent } from '@/uni_modules/sky-nutui/components/sky-nutui/packages/utils/create';
 const { componentName, create } = createComponent('noticebar');
 
-interface StateProps {
-  wrapWidth: number;
-  firstRound: boolean;
-  duration: number;
-  offsetWidth: number;
-  showNoticeBar: boolean;
-  animationClass: string;
 
-  animate: boolean;
-  scrollList: [];
-  distance: number;
-  timer: null;
-  keepAlive: boolean;
-  isCanScroll: null | boolean;
-}
 export default create({
   props: {
     // 滚动方向  across 横向 vertical 纵向
@@ -137,7 +126,7 @@ export default create({
     rightIcon: { type: String, default: '' },
     color: {
       type: String,
-      default: ''
+      default: '#F9911B'
     },
     background: {
       type: String,
@@ -159,7 +148,6 @@ export default create({
   components: {
     ScrollItem: function (props) {
       props.item.props.style = props.style;
-      props.item.key = props.key;
       return h(props.item);
     }
   },
@@ -167,11 +155,13 @@ export default create({
 
   setup(props, { emit, slots }) {
     // console.log('componentName', componentName);
+    
+    const { proxy } = getCurrentInstance()
 
     const wrap = ref<null | HTMLElement>(null);
     const content = ref<null | HTMLElement>(null);
 
-    const state = reactive<StateProps>({
+    const state = reactive({
       wrapWidth: 0,
       firstRound: true,
       duration: 0,
@@ -184,7 +174,8 @@ export default create({
       distance: 0,
       timer: null,
       keepAlive: false,
-      isCanScroll: null
+      isCanScroll: null,
+      id: Math.round(Math.random() * 100000)
     });
 
     const classes = computed(() => {
@@ -196,7 +187,7 @@ export default create({
 
     const isEllipsis = computed(() => {
       if (state.isCanScroll == null) {
-        return props.wrapable;
+        return false && !props.wrapable;
       } else {
         return !state.isCanScroll && !props.wrapable;
       }
@@ -213,10 +204,10 @@ export default create({
     const barStyle = computed(() => {
       let style: {
         [props: string]: any;
-      } = {};
-
-      props.color && (style.color = props.color);
-      props.background && (style.background = props.background);
+      } = {
+        color: props.color,
+        background: props.background
+      };
 
       if (props.direction == 'vertical') {
         style.height = `${props.height}px`;
@@ -231,6 +222,7 @@ export default create({
         transform: `translateX(${state.firstRound ? 0 : state.wrapWidth + 'px'})`
       };
     });
+
     const iconBg = computed(() => {
       let iconBg = '';
       if (props.leftIcon) {
@@ -246,9 +238,8 @@ export default create({
         };
       } else {
         if (state.animate) {
-          let a = ~~(props.height / props.speed / 4);
           styles = {
-            transition: `all ${a == 0 ? ~~(props.height / props.speed) : a}s`,
+            transition: `all ${~~(props.height / props.speed / 4)}s`,
             'margin-top': `-${props.height}px`
           };
         }
@@ -275,25 +266,35 @@ export default create({
         return;
       }
       setTimeout(() => {
-        if (!wrap.value || !content.value) {
-          return;
-        }
-        const wrapWidth = wrap.value.getBoundingClientRect().width;
+        
+        let wrapWidth = 0;
+        let offsetWidth = 0;
+        uni.createSelectorQuery().in(proxy)
+          .select(`.wrap${state.id}`)
+          .boundingClientRect((rect) => {
+            if (rect.width > 0) wrapWidth = rect.width;
+          })
+          .exec();
+        uni.createSelectorQuery().in(proxy)
+          .select(`.content${state.id}`)
+          .boundingClientRect((rect) => {
+              console.log(rect,'rect')
+            if (rect.width > 0) offsetWidth = rect.width;
 
-        const offsetWidth = content.value.getBoundingClientRect().width;
+            state.isCanScroll = props.scrollable == null ? offsetWidth > wrapWidth : props.scrollable;
 
-        state.isCanScroll = props.scrollable == null ? offsetWidth > wrapWidth : props.scrollable;
-        console.log(111, state.isCanScroll);
-        if (state.isCanScroll) {
-          state.wrapWidth = wrapWidth;
-          state.offsetWidth = offsetWidth;
+            if (state.isCanScroll) {
+              state.wrapWidth = wrapWidth;
+              state.offsetWidth = offsetWidth;
 
-          state.duration = offsetWidth / props.speed;
-          state.animationClass = 'play';
-        } else {
-          state.animationClass = '';
-        }
-      }, 0);
+              state.duration = offsetWidth / props.speed;
+              state.animationClass = 'play';
+            } else {
+              state.animationClass = '';
+            }
+          })
+          .exec();
+      }, 100);
     };
     const handleClick = (event: Event) => {
       emit('click', event);
@@ -320,7 +321,7 @@ export default create({
      */
     const startRollEasy = () => {
       showhorseLamp();
-      (state.timer as any) = setInterval(showhorseLamp, ~~((props.height / props.speed / 4) * 1000) + props.standTime);
+      (state.timer as any) = setInterval(showhorseLamp, ~~(props.height / props.speed / 4) * 1000 + props.standTime);
     };
     const showhorseLamp = () => {
       state.animate = true;
@@ -328,7 +329,7 @@ export default create({
         state.scrollList.push(state.scrollList[0]);
         state.scrollList.shift();
         state.animate = false;
-      }, ~~((props.height / props.speed / 4) * 1000));
+      }, ~~(props.height / props.speed / 4) * 1000);
     };
 
     const startRoll = () => {
@@ -363,7 +364,7 @@ export default create({
     onMounted(() => {
       if (props.direction == 'vertical') {
         if (slots.default) {
-          state.scrollList = [].concat(slots.default()[0].children as any);
+          //state.scrollList = [].concat(slots.default()[0].children as any);
         } else {
           state.scrollList = [].concat(props.list as any);
         }
@@ -413,6 +414,7 @@ export default create({
   }
 });
 </script>
+
 <style lang="scss">
 @import './index.scss'
 </style>